ibmveth: Disable tx queue while changing MTU
If the device is running while the MTU is being changed, ibmveth
restarts the device using the newly added restart mechanism.

The pre-existing buffer pools are reset since we are activating new
buffer pools when changing the MTU.

If the restart mechanism fails along the way, the original MTU
configuration is used.

Signed-off-by: Cris Forno <cforno12@linux.vnet.ibm.com>