powerpc/vnic: Extend "failover pending" window
NOTE:	This patch is based on the kernel-source-4.12.14-197.45.1.src.rpm
	for SLES15-SP1 and is intended to fix the customer issue reported
	in bz 187293. Please test to ensure there are no regressions. It
	seems to fix the issue in the simulated customer environment set
	up as described in the bug.

Commit 5a18e1e0c193b introduced the 'failover_pending' state to track
the "failover pending window" - where we wait for the partner to become
ready (after a transport event) before actually attempting to failover.
i.e window is between following two events and ->failover_pending is TRUE.

        a. we get a transport event due to a FAILOVER

        b. later, we get CRQ_INITIALIZED indicating the partner is
           ready  at which point we schedule a FAILOVER reset.

If during this window, we attempt to open (or close) a device, we pretend
that the operation succeded and let the FAILOVER reset complete the
operation.

This is fine, except if the transport event in "a" above occurs after
open() checks whether a failover is pending. i.e the transport event
occurs _during_ open, at which time we fail the open. This can cause
the boot scripts to leave the interface down requiring administrator
to manually bring up the device.

This fix "extends" the failover pending window till we are _actually_
ready to perform the failover reset (i.e until after we get the RTNL
lock). Since open() holds the RTNL lock, we can be sure that we either
finish the open or if the open() fails due to the failover window, we
can again pretend that open is done and let the failover complete it.
Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.ibm.com>